Simplify odl scenarios
So far we were building the repo url in the odl scenarios. Moving that logic
to ODL role probably makes more sense as it can be reused by all scenarios.

This patch passes the odl_version to OSA. That variable will be processed by
the ODL role. That role will build the repo url

deploy-scenario:os-odl-nofeature
installer-type:osa
Change-Id: I0c9da069238348af6d00d8422ca478cb4be4cfad
Signed-off-by: Manuel Buil <mbuil@suse.com>