xci: kubespray: Fix failures to due task checks
(this commit fixes many things because they all need to be submitted
together to unblock the jobs)

Commit 9e1d3d6e62abf5d0da26a296bcd235f37a54d9c6 ("xci: playbooks: Fixes
various ansible-lint warnings") broke public key authentication from
localhost to the OPNFV host because the localhost pubkey was not
appended in the authorized_keys file. The reason for that was that the
task was skipped due to the 'creates' parameter. This is now fixed, by
dropping the check since we always need to append the localhost pubkey.

This is only a temporary solution until we modify kubespray to use the
common file for managing the SSH keys.

This also makes the final 'kubectl' move to /usr/local/bin non-fatal
since future kubespray releases put it there already.

The same commit also broke the k8s-cluster.yml overrides. This is
because the file was never copied across due to the task conditional
being wrong. As such, we fix the conditional to check for the correct
file.

Change-Id: I9cfb29eba50c7fea9df29581ebb015163b8a9754
Signed-off-by: Markos Chandras <mchandras@suse.de>