Update patch set 2
Patch Set 2:

> Yes, it is not only to explain and take the virtual pod as an
 > example, but also to propose to use this file as config source for
 > the installers and shared roles.
 > 
 > Those files are extracted from prototypes/infra_manager, and they
 > are used today to deploy a virtual pod with the proto

Ok. But note that we will not always deploy 5 nodes. For example, the most common deployment uses the mini flavor which has 1 controller and 1 compute. I don't know how flexible is the idf/pdf processing but I guess it should just ignore the 2 controllers and the extra compute in case there is a mini deployment, or?

Patch-set: 2