Update patch set 2
Patch Set 2: Code-Review-1

(16 comments)

this is what I have been looking for and I believe this change does the things in right way if I am guessing/following things up correctly as written below.

the first thing is to switch to dynamic inventory generation, followed by the introduction of pdf/idf. and later, switching to generating the osa config files dynamically.

Another question I have is whether the jobs verified this patch used the dynamically generated inventory or the static ones? If the dynamically generated ones are used, it is even greater since I don't see much else to be done for this change except looking at the comments inline and adressing them.

Patch-set: 2
Label: Code-Review=-1