Update patch set 4
Patch Set 4:

30 lines of code for now, to create a clean json structure... in bash... maybe
... but how much when we will introduce PDF/IDF parsing ? adding the basic parameters in host_vars ? and this is the goal of this patch.

For me, the main questions are:
- do we agree on moving inventories in the common parts (role or not) ?
- do we agree on having 1 inventory by type of installer until we can converge smoothly to a single one ?
- do we agree on a "non installer specific naming" of node functions or do we use OpenStack naming for all ?

Again, writing a dynamic inventory is not a technical problem, but it is a rewriting of something that is working, that I can enhance easily by pasting existing code to parse PDF/IDF. It is not possible to be quick and rewrite everything just to gain 100 lines of codes (and it seems that baremetal introduction is some kind of priority).

If there is no consensus on that proposition until the end of the week I will rewrite a patch with dynamic inventory.

Patch-set: 4
Reviewer: Gerrit User 361 <361@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0