State which project scenarios belong to
A new case has been identified where the current promotion logic
fails to cover.

If a change is sent to releng-xci and if the developer who proposed the change
states that it impacts an external scenario by specifying it in commit message
using deploy-scenario string, CI deploys and tests that change as part of verify
jobs using the specified scenario rather than one of the generic scenarios.

When the change gets merged and if the deployment and testing end successfully,
the promotion job uploads the properties file which contains metadata such as
SHAs of releng-xci and the project repo which the scenario belongs to.
An example can be seen on [1].

But the promotion script/logic doesn't cover this case and puts the SHA of
releng-xci as SCENARIO_SHA into metadata file since it purely looks at the
Gerrit Project where the change gets merged in rather than the actual scenario
which might be impacted due to change, resulting in the storage of incorrect
promotion metadata. An example where the logic fails can be seen on [2] which
failed thanks to Jenkins slave issues which prevented us to promote os-odl-sfc
with incorrect metadata.

By adding which project a certain scenario belongs to into opnfv scenario
requirements file, we can use the project name which the scenario belongs to
for extracting SHA rather than using Gerrit Project where the change is proposed
to.

Not using src is on purpose since scenario names don't always match to project name
such as os-odl-bgpvpn belongs to sdnvpn project. Or projects might have multiple
scenarios again sdnvpn will have another scenario named os-nosdn-bgpvpn.

[1] http://artifacts.opnfv.org/xci/pipeline/merge/os-odl-bgpvpn.properties
[2] https://build.opnfv.org/ci/job/xci-merge-virtual-master/62/

Change-Id: I5301630a71446df8668f16c81bf26d81bcea8031
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>