Update patch set 8
Patch Set 8:

After the import of the XCI VMs, The openstack services (neutron-server, DHCP agent, Metering agent, Metadata agent, L3 agent, Open vSwitch agent, nova-conductor and nova-comute) are started in random order which causes neutron and nova is not able to find DHCP agent and compute accordingly to bring up the tenant VM. 
Once these services are rebooted in certain order, things started working fine. So we need to follow required post import steps to get the system in working state (like rebooting services in proper order). Do we have ansible task in OSA to perform this ? 
If there are such things exists already, then we can integrate it with this script depending on the import scenario/flavor.

Patch-set: 8