Start documenting CI Loops and Confidence Loops applied by XCI
This document is created to start the conversation around determining
the promotion criterias (test scopes) for each of the current and upcoming
CI Loops XCI uses.

The expectation with this document is to update it collaboratively with
scenario owners, projects, XCI Team, and Test Projects to find the right
level of testing as much as possible.

The information this document contains and the criterias set in it will be
superseded by the CD-Based Release model becomes available.

Change-Id: I7b4afd87a4f7694eb78c996e7870b8dc1e833fbf
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>