Start documenting CI Loops and Confidence Loops applied by XCI
Please note that the document you are about to review contains about the
current CI Loops and promotion criterias in order to record the current
state of affairs.

Once this change gets merged, a new change will be proposed and that change
will be the one where the criterias will be worked on.

This document is created to start the conversation around determining
the promotion criterias (test scopes) for each of the current and upcoming
CI Loops XCI uses.

The expectation with this document is to update it collaboratively with
scenario owners, projects, XCI Team, Test Projects, and Release Management
to find the necessary level of testing as much as possible.

The information this document contains and the criterias set in it will be
superseded by the CD-Based Release model once it becomes available.

Change-Id: I7b4afd87a4f7694eb78c996e7870b8dc1e833fbf
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>