Update patch set 4
Patch Set 4: Code-Review+2

I think this could be ok for the time being, even if it does not follow ansible best practices and thus creates some technical debt which we should address afterwards. The baremetal support involves a lot of patches which are difficult to review as a whole, so I'd prefer merging something which is acceptable and works. We can improve it later with smaller patches

Patch-set: 4
Reviewer: Gerrit User 2091 <2091@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2