Pin os-odl-bgpvpn to project internal tag 6.0.0
It is important to test and pin the versions tagged by the projects
to ensure we have a single version of XCI with pinned versions of the
scenarios for all the projects are recorded.

Same exercise will be done for all the scenarios and then scenarios will
be unpinned from the tags to let the development continue.

deploy-scenario:os-odl-bgpvpn
installer-type:osa

Change-Id: I9350bc53fa87836ad36259b259edf506dd99790a
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>