Update patch set 1
Patch Set 1:

> Do the periodic promotion levels map to specific release types? 
 > For example, if we establish a "developer" release and an "end
 > user" release, which promotion levels would those map to?

Daily and weekly doesn't actually represent how frequently these loops are run or testing done. They should be seen as different CI Loops, each testing different aspects of the composed platform. The frequency of these loops are determined by the projects' themselves; if a project is slow developing features/integrating upstream, they might run "daily" once a week since they might not have any version that is qualified to be tested on baremetal on a daily basis. At the same time, a project is doing development actively and their versions pass post-merge loops consistently, they might get more than 1 "daily" runs.

Release types also depend on the projects' themselves, especially the developer release. If projects are actively developing features and fulfilling the criterias for "daily" for example, they can apply a project internal tag in the form of x.y.z and this can be seen as developer release.

For "end user" release; this also depends on projects. For each release cycle, there should be 1..n release candidates tagged with internal project tag which can then be promoted as official OPNFV/"end user" release, getting the umbrella tag opnfv-x.y.z. If there is no internal project tag for a long time, the criteria might need to be reviewed or the project is dropped from the release.

One problem with what is documented here is the naming of loops and corresponding confidence levels. It doesn't really tell much about what is tested in certain CI loop. For weekly, we can and should perhaps use the name "long duration testing (ldt)" so it can give an idea about what that loop does. But I'm not sure what to call the daily, telling more about what is tested there. Another problem with the naming is that it gives sense that we will run things periodically. In XCI, the intention is to run things when things happen; a new change got merged to master, let's run post-merge loop to see if is it good enough to be tested on baremetal. Or a version of a certain scenario passed post-merge loop, it should be tested on baremetal which uses tougher test criteria. So CI loops used XCI will mostly use events rather than running things based on timer/polling.

Patch-set: 1