xci: playbooks: Simplify scenario overrides mechanism
Now that releng-xci does not have any internal scenarios anymore, we can
simply the way we plug scenarios to XCI. We keep cloning scenarios in
the cache directory but we also allow users to pass a special
'xci_scenarios_overrides' variable similar to the
'opnfv_scenario_requirements.yml' file to allow them to override
specific keys of the dictionaries. This can be used by Jenkins to test
incoming changes by altering the defaults values. For example, when
we can to test a Gerrit change we can create a file called foobar.yml
with the following contents

xci_scenarios_overrides:
  - scenario:
      version: 9fa5400012e02408b416eb7bd26307e7d179abce
      refspec: refs/changes/79/55879/7

and load it using

XCI_ANSIBLE_PARAMS="-e @foobar.yml"

This will make XCI checkout the refspec that we want instead of the
default version.

Change-Id: Ida49910b4c79b843582ddd70c00063fe9bc90dd4
Signed-off-by: Markos Chandras <mchandras@suse.de>