Update patch set 4
Patch Set 4:

> (1 comment)

OK so this is not quite meant to be used like you describe. Look at https://gerrit.opnfv.org/gerrit/#/c/56669/

The idea is that Jenkins can create such file after it determines the scenario that it's about to be tested and feed that information to XCI. Does that answer the question? This file is used to create scenario patches against XCI not the other way around.

Patch-set: 4