Update patch set 1
Patch Set 1:

> This repo is about openstack ansible installer providing XCI in a
 > sandbox environment.  I'm not sure why its called releng-xci, when
 > xci can encompass other installers who do not live in this repo. 
 > Even the install command is called "xci-deploy.sh", which runs
 > openstack ansible and bifrost.  I'm pretty against this repo's name
 > and it housing any documentation that contains references which
 > imply XCI is openstack ansible.


Thanks for the feedback.

This change and the other changes this one depends on are xci specific and only targeted for the projects that are onboarding to xci. (xci != cd release process)

As explicitly stated in commit message and the initial mail sent to the community, this will be replaced by CD release model once it becomes available so the comment you provided is not valid in this context.

Patch-set: 1
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0