Use PDF to set up the XCI VMs
This patch is doing the first work item of the spec:

https://github.com/opnfv/releng-xci/blob/master/docs/specs/infra_manager.rst

It creates the required VMs by XCI to afterwards deploy the VIM. It does that
by reading the pdf provided by the user. Please read the README for more
information. Important assumptions (covered in README):

- It is currently assumed that the OS for the VM will be installed in the first
disk of the node described by the pdf

- It is assumed that the opnfv VM characteristics are not described in the pdf
but in a similar document called opnfv_vm.yml

- All references to csv from bifrost-create-vm-nodes were removed

Apart from that, to ease the review, here are some points:
* The opnfv VM is created first and then the rest. All use the playbook:
create_vm.yml

* opnfv VM is provisioned first and then, from that VM, the rest are
provisioned using bifrost (ironic running in the opnfv VM)

Change-Id: I46a85284e4ce7df21cbf66f66619b35f74251e68
Signed-off-by: Manuel Buil <mbuil@suse.com>
Signed-off-by: Markos Chandras <mchandras@suse.de>