DNM: xci: Bump OSA SHAs, pin ceph-ansible and Ansible
deploy.sh has a dirty hack to get around the recently appeared
problem with not being able to reach to the target nodes. Due to this,
fact gathering fails for containers and the deployment is interrupted.
It is put there just to see how far the deployment goes with the
shabump and make it possible for further troubleshooting. See
https://build.opnfv.org/ci/job/xci-verify-ubuntu-deploy-virtual-master/1733/console
for the actual problem with the broken fact gathering when target
nodes are not resolved correctly. This might be due to the use of
dynamic inventory and needs to be properly fixed.

This change
- bumps OSA SHA to 54e21afb59152ec276f725dce8479f84db7b395f
- pins ceph-ansible role to 78c1f1938f9365809cc5e7d09cb77566c7045e3b
- pins Ansible to 2.4.4.0

Change-Id: Icc14e3e794b489dafd78b426c54051a3732ccb1a
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>