[WIP] - Baremetal support
Steps (and potential patch split):

1 - Connect the interface which will receive pxe traffic to
the bridge for opnfv vm. That interface is defined in
PXE_INTERFACE and its value goes into network_interface_pxe

2 - Do the same as #1 with a bridge that provides the mgmt traffic
to the opnfv VM (at one point, deploy.sh moves to mgmt network to
do the Ansible config, so we need it). The interface is defined as
MGMT_INTERFACE.

3 - Create the baremetal yaml file using idf and pdf through
the playbook baremetalhosttojson.yml. This file is consumed by ironic

4 - Use bifrost-create-dib-image to create a Fedora image.
Ericsson pod2 only works with that one. To do so we need

download_ipa=false (we don't want tinylinux image)
create_ipa_image=true

5 - Configure static ip addresses based on the mac address for the
dnsmasq config used by ironic:
 * Generate baremetalstaticips
 * inventory_dhcp=true
 * Store baremetalstaticips in /etc/dnsmasq.d/bifrost.dhcp-hosts.d/
As libvirt is not providing the DHCP service anymore

6 - Change the network config so that it uses the physical interfaces
for the baremetal blades

7 - Change the dynamic_inventory so that it adds MACs and VLAN. That
helps a lot when doing the network config in #6

8 - As opnfv_vm has now two interfaces (pxe and mgmt), we must configure
the default route correctly, otherwise ipmi connectivity gets broken

9 - Improve the blades connectivity check. Blades are not ready with the
official OS (not IPA) when the code moves to configure the opnfv VM network.
That creates glitches in the ipmi connecitivity which makes ironic fail

10 - Correct the idf and pdf of ericsson-pod2 because it was wrong. Complete
the opnfv_vm file too

11 - Provide a new network config for opnfv VM when doing baremetal. Make sure
the ip is 0 in PXE and MGMT interface when restarting the network

12 - Create a idf for opnfv-vm so that we can reuse as much code as possible in
virtual and baremetal deployment. The reason is that we need a way to know what
are we doing with each interface (pxe, mgmt...)

Change-Id: I08f5d921f2adcc924ee1e0f6d945a6c9b2a24699
Signed-off-by: Manuel Buil <mbuil@suse.com>