Disable the flavor aio and set default flavor to mini
The flavor all-in-one (aio) hasn't been maintained well and
the stability of it is in question. This change disables the
flavor aio and sets the default flavor to mini.

It is important for XCI to enable the aio back in order to help
users who want to try XCI but don't have big enough machine to
go for bigger flavors. However, we need to verify all flavors
for acceptable user experience so the flavor can only be enabled
back once this happens.

Change-Id: I053c8fec78876dea71439f1f1bd737a105254ad4
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>