WIP: Integrate os-nosdn-osm scenario
This change integrates OSM into XCI using a newly created scenario
os-nosdn-osm.

The biggest impact this change has is the addition of the new playbook
named post-deployment.yml which enables running post-deployment tasks
that are defined as part of the scenario. If post-deployment.yml exists
in scenario role, tasks from it are run as part of post-deployment. If
the file does not exist, it is skipped.

installer-type:osa
deploy-scenario:os-nosdn-osm

Change-Id: I16780abffca39699eb2f38f662479f0e4d551504
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>