WIP: Integrate os-nosdn-osm scenario
This change integrates os-nosdn-osm scenario.

The installation of OSM is done as post-deployment. This is achieved by the
addition of the new playbook named post-deployment.yml in scenario role folder.

This mechanism is available for all OpenStack scenarios. If anything needs to be
done for a specific scenario as part of the post-deployment, it can be achieved
by creating playbook post-deployment.yml.

If post-deployment.yml exists in scenario role, the framework will run it once
once the regular deployment playbooks are run successfully. If the file does
not exist, it will be skipped. The location of the post-deployment.yml is
    releng-xci-scenarios/<scenario>/role/<scenario>/tasks/post-deployment.yml

This is only implemented for OpenStack scenarios currently and similar mechanism
is needed for K8S scenarios as well which will be implemented in a separate
change.

installer-type:osa
deploy-scenario:os-nosdn-osm

Change-Id: I16780abffca39699eb2f38f662479f0e4d551504
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>