Pin functest-healthcheck docker image to known sha
Functest changes have significant impact which blocks everything in XCI so this
change pins the image to a known sha to get the original set of healthcheck
testcases until the impacts are analysed and concerns are raised to Functest
and the wider OPNFV Community and addressed based on community consensus.

Pinned version of functest-healthcheck contains the test cases below.
- connection_check
- api_check
- snaps_health_check

deploy-scenario:os-nosdn-nofeature
installer-type:osa

Change-Id: Ic9222af8c27e58491b7b60a7504df9d792b5e753
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>