Enable ability to run post-deployment tasks for OpenStack scenarios
This change makes it possible to run post-deployment tasks for OpenStack
scenarios.

This is achieved by the addition of the new playbook named post-deployment.yml.
If post-deployment.yml exists in scenario role, the framework will run it once
once the regular deployment playbooks are run successfully. If the file does
not exist, it will be skipped. The location of the post-deployment.yml is
    releng-xci-scenarios/<scenario>/role/<scenario>/tasks/post-deployment.yml

This is only implemented for OpenStack scenarios currently and similar mechanism
is needed for K8S scenarios as well which will be implemented in a separate
change.

installer-type:osa
deploy-scenario:os-nosdn-osm

Change-Id: I3c67802160a251d2c27d308c50e59825b5d70032
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>