[baremetal] idf for opnfv_vm and update the pdf
Physical hardware PODs provide a pdf and a idf to describe hardware and other
information (e.g. what is the purpose for each interface). To reuse the
same come for opnfv vm, we need an idf for the opnfv vm.

As we add an idf, we should update dynamic_inventory too

The pdf of opnfv vm has four interfaces but we only need two for admin and
pxe interfaces

The virtual idf is also updated to include the pxe network

Change-Id: I930728474631fc214e4a9adc8581e0c16d230176
Signed-off-by: Manuel Buil <mbuil@suse.com>