[baremetal] idf for opnfv_vm and update the pdf
Physical hardware PODs provide a pdf and a idf to describe hardware and other
information (e.g. what is the purpose for each interface). To reuse the
same come for opnfv vm, we need an idf for the opnfv vm.

As we add an idf, we should update dynamic_inventory too

The pdf of opnfv vm has four interfaces but we only need two for mgmt and
admin interfaces

We add the physical interfaces that are used for mgmt and admin in the
deployment_host vars (they are fetched from idf)

As we are replacing the opnfv_vm.yml file, we remove it

Change-Id: I930728474631fc214e4a9adc8581e0c16d230176
Signed-off-by: Manuel Buil <mbuil@suse.com>