[Baremetal] Add important vars for baremetal
BAREMETAL: It switches between parts of the code which are specific for
baremetal or non-baremetal

PXE_INTERFACE: Specifies what interface in the jumphost will be used
for PXE in the baremetal deployment

MGMT_INTERFACE: Specifies what interface in the jumphost is used for
mgmt/admin traffic in the baremetal deployment

The interfaces could be fetched through idf/pdf but it might become
quite difficult because the jumphost networking might be complex. That
is why, I decided for the time being to use a variable. Note that the
only way to get the interface through idf/pdf is by using the  mac
address and the same mac might be shared among several network devices
(e.g. eth0, eth0.10, eth0.20, br-whatever if it has one of the eth0
interfaces)

Change-Id: I8e6171f4f21db7f814a472e6ed1bacb30220b4ec
Signed-off-by: Manuel Buil <mbuil@suse.com>