[Baremetal] Two interfaces for opnfv vm
opnfv vm requires connectivity to two physical interfaces of the host.
These interfaces are:

1 - The one where DHCP requests will arrive from blades to do PXE boot
2 - The one which connects to the mgmt of the blades to do the ansible
configuration

To achive this, it is required:

1 - Two libvirt networks that connect to two different linux bridges.
The important physical interfaces are connected to them. The interfaces
name are fetched from the idf

2 - For PXE to work, we create a file called baremetalstaticips that
has the mapping between mac address from blades and its ip. That file
is moved into the dnsmasq config directory

3 - baremetalhoststojson.yml fills the json config file fed to ironic
so that it knows how to boot the blades. In the baremetal case, the
create_vm.yml playbook will only create opnfv vm. The variable
vms_to_create holds that information. The variable baremetal_nodes
specifies the physical nodes (empty when non baremetal deployments)

4 - Two templates representing the new libvirt networks (net-mgmt.xml.j2
and net-admin.xml.j2)

5 - Two interfaces defined in vm.xml.j2

6 - The create-vm-nodes role gets a name change to reflect what it does.
Now the name is create-nodes. xci-create-virtual changes to
xci-setup-nodes.

Change-Id: I0e788db1deb50769c183b71524a68ac0b925f8aa
Signed-off-by: Manuel Buil <mbuil@suse.com>