Update patch set 29
Patch Set 29:

I faced the same error when I tried deploying using this patch ubuntu. 

The reason of the failure is /etc/openstack_deploy/user_variables_xci.yml not having anything but license header. I don't know how it worked until now but adding something in it makes it pass that point. Later on it failed with glance for me.

Patch-set: 29