Update patch set 1
Patch Set 1: Code-Review+2

not related to this patch but to vbmc: i had trouble with its 1.4.0 version on opensuse while the created vms are attempted to be added to vbmc. the error is "the vm is already running" even though it is newly added and not running.

should we pin it to 1.3.0? this is the one we have been using.

Patch-set: 1
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2