xci: roles: create-vm-nodes: Always provision VM nodes
The role used to get the output of 'virsh list --all' to determine how
many VMs are present and shut off in the system. This takes *all* VMs
on the system into consideration so it may skip creation of some or all
of the XCI VMs if we happen to have other VMs present. We can improve
the situation by simply dropping this check and always provision the
VMs we want. If the VM is already present, then the module will simply
do a sanity check of its configuration. This allows XCI to run alongside
other VMs.

Change-Id: I54255a1959509671c0305f48f23a55b6e900684f
Signed-off-by: Markos Chandras <mchandras@suse.de>