Rename functest-prepare role to prepare-tests
This change renames functest prepare-tole to prepare tests and makes other
adjustments to the role to move common test preparation steps to its own
script so we can prepare for Functest and Yardstick at one go.

Similar things are required to be prepared for running Functest and Yardstick
such as
- installed packages
- external network creation
- creation of run-functest.sh and run-yardstick scripts from templates
- preparation of environment variables

This change will fail verification until the changes are reflected to
verify jobs which will be done separately and change will be reverified.

Change-Id: Id1020d3e61abd3f087863c06a132c5021339d655
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>